Added a lot of new functionality.
Added the material library.
Added a dialog in order to add more exercises to the lists. This dialog contains a form that makes some validations.
Modified some properties from the designer component.
Added a pipe in order to show the difficulty in more than one color.
Added  the result items in order to give it a different fnctionality.